+++ previous-versions/accounting/src/codataccounting/models/operations/create_transfer.py
@@ -13,7 +13,9 @@ from typing import Optional
 @dataclasses.dataclass
 class CreateTransferRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     transfer: Optional[shared_transfer.Transfer] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
 
@@ -23,11 +25,14 @@ class CreateTransferRequest:
 @dataclasses.dataclass
 class CreateTransferResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     create_transfer_response: Optional[shared_createtransferresponse.CreateTransferResponse] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""