+++ previous-versions/accounting/src/codataccounting/models/operations/download_invoice_pdf.py
@@ -10,8 +10,9 @@ from typing import Optional
 @dataclasses.dataclass
 class DownloadInvoicePdfRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     invoice_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'invoiceId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an invoice"""
+    r"""Unique identifier for an invoice."""
     
 
 
@@ -20,9 +21,12 @@ class DownloadInvoicePdfRequest:
 @dataclasses.dataclass
 class DownloadInvoicePdfResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     data: Optional[bytes] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""