+++ previous-versions/accounting/src/codataccounting/models/operations/get_account.py
@@ -12,8 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetAccountRequest:
     account_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'accountId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an account"""
+    r"""Unique identifier for an account."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     
 
 
@@ -22,11 +23,14 @@ class GetAccountRequest:
 @dataclasses.dataclass
 class GetAccountResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     account: Optional[shared_account.Account] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""