+++ previous-versions/accounting/src/codataccounting/models/operations/get_invoice.py
@@ -12,8 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetInvoiceRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     invoice_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'invoiceId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an invoice"""
+    r"""Unique identifier for an invoice."""
     
 
 
@@ -22,11 +23,14 @@ class GetInvoiceRequest:
 @dataclasses.dataclass
 class GetInvoiceResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     invoice: Optional[shared_invoice.Invoice] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""