+++ previous-versions/accounting/src/codataccounting/models/operations/get_payment.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetPaymentRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     payment_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'paymentId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a payment."""
     
 
 
@@ -21,11 +23,14 @@ class GetPaymentRequest:
 @dataclasses.dataclass
 class GetPaymentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     payment: Optional[shared_payment.Payment] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""