+++ previous-versions/accounting/src/codataccounting/models/operations/upload_direct_income_attachment.py
@@ -20,8 +20,11 @@ class UploadDirectIncomeAttachmentRequestBody:
 @dataclasses.dataclass
 class UploadDirectIncomeAttachmentRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     direct_income_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'directIncomeId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a direct income."""
     request_body: Optional[UploadDirectIncomeAttachmentRequestBody] = dataclasses.field(default=None, metadata={'multipart_form': { 'file': True }, 'request': { 'media_type': 'multipart/form-data' }})
     
 
@@ -31,9 +34,12 @@ class UploadDirectIncomeAttachmentRequest:
 @dataclasses.dataclass
 class UploadDirectIncomeAttachmentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""