+++ previous-versions/accounting/src/codataccounting/models/shared/bill.py
@@ -109,7 +109,9 @@ class BillSupplementalData:
 @dataclasses.dataclass
 class BillWithholdingTax:
     amount: Decimal = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('amount'), 'encoder': utils.decimalencoder(False, False), 'decoder': utils.decimaldecoder }})
+    r"""Amount of tax withheld."""
     name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name') }})
+    r"""Name assigned to withheld tax."""