+++ previous-versions/accounting/src/codataccounting/models/shared/createbanktransactions.py
@@ -13,6 +13,7 @@ from typing import Optional
 @dataclasses.dataclass
 class CreateBankTransactions:
     account_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('accountId'), 'exclude': lambda f: f is None }})
+    r"""Unique identifier for a bank account."""
     transactions: Optional[list[shared_createbankaccounttransaction.CreateBankAccountTransaction]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('transactions'), 'exclude': lambda f: f is None }})